Pass connector_id to /auth on completion of device auth flow + add per-client default connector_id (#326)
* Pass connector_id to /auth on completion of device auth flow

This allows users to select a connector when initiating device auth.
They simply need to append "connector_id" to the query string of
"/device"  alongside "user_code". The form will submit the chosen
connector to Dex which will add the connector to  "/auth" to skip the
webpage for selecting a connector.

* Add default connector IDs in /auth for specific client IDs

If a connector_id query param isn't provided, Dex will ask the user to
select one. Clients can specify a default connector ID to skip this
manual user step. This change will get `client_id` from the query string
and check if there's a default connector.

DEFAULT_CONNECTORS_FILE will provide the path of a yaml file of this
format:
```
clientid1: connector1
clientid2: connector2
```